Moved relevant tsx files from src/app to src/app/home-page
