+++ src/app/home-page/string_functions.tsx
